+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Évolutions
 - Permettre la connection à Passtrans #1024
+- Permettre l'export du carnet d'adresse d'une collectivité #1033
 
 ## 5.0.10 - 2023-05-17